+++ providers/darwin/process_darwin_test.go
@@ -15,14 +15,16 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//go:build (amd64 && cgo) || (arm64 && cgo)
-// +build amd64,cgo arm64,cgo
+//go:build amd64 || arm64
+// +build amd64 arm64
 
 package darwin
 
 import (
+	"errors"
 	"os"
 	"os/exec"
+	"syscall"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -93,4 +95,48 @@ func TestProcessEnvironmentInternal(t *testing.T) {
 	assert.Equal(t, "", value)
 
 	assert.Equal(t, "FOO", p.env[fooValueEnvVar])
+}
+
+func TestProcesses(t *testing.T) {
+	var s darwinSystem
+	processes, err := s.Processes()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	var count int
+	for _, proc := range processes {
+		processInfo, err := proc.Info()
+		switch {
+		// Ignore processes that no longer exist or that cannot be accessed.
+		case errors.Is(err, syscall.ESRCH),
+			errors.Is(err, syscall.EPERM),
+			errors.Is(err, syscall.EINVAL):
+			continue
+		case err != nil:
+			t.Fatalf("failed to get process info for PID=%d: %v", proc.PID(), err)
+		default:
+			count++
+		}
+
+		if processInfo.PID == 0 {
+			t.Fatalf("empty pid in %#v", processInfo)
+		}
+
+		if processInfo.Exe == "" {
+			t.Fatalf("empty exec in %#v", processInfo)
+		}
+
+		u, err := proc.User()
+		require.NoError(t, err)
+
+		require.NotEmpty(t, u.UID)
+		require.NotEmpty(t, u.EUID)
+		require.NotEmpty(t, u.SUID)
+		require.NotEmpty(t, u.GID)
+		require.NotEmpty(t, u.EGID)
+		require.NotEmpty(t, u.SGID)
+	}
+
+	assert.NotZero(t, count, "failed to get process info for any processes")
 }