+++ types/host.go
@@ -20,10 +20,13 @@ package types
 import "time"
 
 // Host is the interface that wraps methods for returning Host stats
+// It may return partial information if the provider
+// implementation is unable to collect all of the necessary data.
 type Host interface {
 	CPUTimer
 	Info() HostInfo
 	Memory() (*HostMemoryInfo, error)
+	FQDN() (string, error)
 }
 
 // NetworkCounters represents network stats from /proc/net
@@ -96,7 +99,7 @@ type OSInfo struct {
 // LoadAverage is the interface that wraps the LoadAverage method.
 // LoadAverage returns load info on the host
 type LoadAverage interface {
-	LoadAverage() LoadAverageInfo
+	LoadAverage() (*LoadAverageInfo, error)
 }
 
 // LoadAverageInfo contains load statistics