+++ CHANGELOG.md
@@ -9,17 +9,47 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Replace pkg/errors with Go 1.13 native errors #123.
-
 ### Changed
-
-- Remove custom sysctl implementation and partial cgo requirement 
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+## [1.10.0]
+
+### Added
+
+- Add OS family mappings for `opensuse-leap` and `opensuse-tumbleweed`. [#146](https://github.com/elastic/go-sysinfo/pull/146)
+- Add FQDN to host info. [#144](https://github.com/elastic/go-sysinfo/pull/144)
+- Return FQDN and error via `FQDN()` method on `Host` interface. [#159](https://github.com/elastic/go-sysinfo/pull/159)
+- Bump golang.org/x/net from 0.0.0-20220722155237-a158d28d115b to 0.7.0 [#153](https://github.com/elastic/go-sysinfo/pull/153)
+- Bump github.com/docker/docker from 20.10.22+incompatible to 20.10.24+incompatible [#160](https://github.com/elastic/go-sysinfo/pull/160)
+
+### Changed
+
+- Requires Go 1.18+ [#144](https://github.com/elastic/go-sysinfo/pull/144)
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+- On darwin without CGO `process.Info()` could fail, but would not return the error. [#150](https://github.com/elastic/go-sysinfo/pull/150)
+
+## [1.9.0]
+
+### Added
+
+- Replace pkg/errors with Go 1.13 native errors. [#123](https://github.com/elastic/go-sysinfo/pull/123)
+- Add OS family mappings for `rocky`, `openEuler`, and `almalinux`. [#143](https://github.com/elastic/go-sysinfo/pull/143)
+
+### Changed
+
+- Remove custom sysctl implementation and partial cgo requirement. [#135](https://github.com/elastic/go-sysinfo/pull/135)
+- Changes on the `Host` and `LoadAverage` interfaces, now implemented by default on Linux and Darwin platforms. [#140](https://github.com/elastic/go-sysinfo/pull/140)
 
 ## [1.8.1]
 
@@ -139,7 +169,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Changed the host containerized check to reduce false positives. [#42](https://github.com/elastic/go-sysinfo/pull/42) [#43](https://github.com/elastic/go-sysinfo/pull/43)
 
-[Unreleased]: https://github.com/elastic/go-sysinfo/compare/v1.8.1...HEAD
+[Unreleased]: https://github.com/elastic/go-sysinfo/compare/v1.10.0...HEAD
+[1.10.0]: https://github.com/elastic/go-sysinfo/releases/tag/v1.10.0
+[1.9.0]: https://github.com/elastic/go-sysinfo/releases/tag/v1.9.0
 [1.8.1]: https://github.com/elastic/go-sysinfo/releases/tag/v1.8.1
 [1.8.0]: https://github.com/elastic/go-sysinfo/releases/tag/v1.8.0
 [1.7.1]: https://github.com/elastic/go-sysinfo/releases/tag/v1.7.1